Implemented graph and graph element copying semantics.
Graphs and graph elements have rudimentary 'constructors' and 'copy
constructors' that properly set up the prototypes. The logic for this could use
some streamlining (to the tune of Backbone's extend helper methods) and
eventually it should probably get broken out of the graph model code.The status
classes for labeling graph elements are now all namespaced with the 'status-'
prefix (but note that in order to set or check a status, you must use the
prefix-free version, e.g. addStatus('foo') to add class 'status-foo').